Standardized test location and namespaces
