+++ nbproject/project.properties
@@ -9,7 +9,9 @@ auxiliary.org-netbeans-modules-php-phpunit.phpUnit_2e_enabled=false
 auxiliary.org-netbeans-modules-php-phpunit.phpUnit_2e_path=
 auxiliary.org-netbeans-modules-php-phpunit.test_2e_groups_2e_ask=false
 auxiliary.org-netbeans-modules-php-phpunit.test_2e_run_2e_all=false
+auxiliary.org-netbeans-modules-php-phpunit.test_2e_run_2e_phpunit_2e_only=false
 file.reference.test-functional=test/functional
+file.reference.test-unit=test/unit
 file.reference.validator-interface-vendor=vendor
 include.path=\
     ${php.global.include.path}:\
@@ -19,6 +21,6 @@ source.encoding=UTF-8
 src.dir=src
 tags.asp=false
 tags.short=false
-test.src.dir=${file.reference.test-functional}
+test.src.dir=${file.reference.test-unit}
 testing.providers=PhpUnit
 web.root=.
